# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Lectio
  version: 0.0.0
  summary: Parse a file into pages, blocks, tables, and fields.
  description: |
    A parse turns one file into one document. It is asynchronous: a submit
    answers at once, and the document is read as it is built, one page at a
    time or whole.

    What a parse does is fixed at submit by a small set of options, each
    independent of the others: which pages, which reader, in which class
    and with which priority, how many failed pages are tolerated, and by
    when. What a caller sees of the result is chosen when it is read: the
    format, the pages, how tables are written. Reading a result in another
    form never runs the parse again. Chunking is a view of a document and
    extraction is a request against one, so neither is a submit option.

    A body with a field this document does not describe is refused with
    `unknown_field`. A route marked `x-lectio-status: planned` is part of the
    contract and answers `501 not_implemented` until it is built.

    Every error has the same shape: one `code` to branch on, one `message`
    for a person, and `details` for a developer. Any operation may answer
    `401`, `403`, `429`, `500` and `503` in that shape; an operation lists
    the other statuses it answers. An address that is not routed is `404
    not_found`, and one that is routed under another method is `405
    method_not_allowed`.

...
    `document_not_ready` and `details.retryable: true`: the address is
...
servers:
...
security:
  - bearer: []
...
paths:
  /files:
    post:
      tags: [files]
      operationId: createFile
      summary: Upload a file
      description: |
        The body is the file itself, with its media type in `Content-Type`,
        or `multipart/form-data` with one part named `file`. It is stored
        while it is hashed and never held whole in memory. Uploading bytes
        the caller already has returns the existing file with `200`.
      parameters:
        - name: name
          in: query
          description: The file's name, used only to help tell its type when the content does not.
          schema: { type: string, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200": { description: The caller already has these bytes., content: { application/json: { schema: { $ref: "#/components/schemas/File" } } } }
        "201": { description: Stored., content: { application/json: { schema: { $ref: "#/components/schemas/File" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /files/{file}:
    parameters:
      - $ref: "#/components/parameters/file"
    get:
      tags: [files]
      operationId: getFile
      summary: Read a file's metadata
      responses:
        "200": { description: The file., content: { application/json: { schema: { $ref: "#/components/schemas/File" } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [files]
      operationId: deleteFile
      summary: Delete a file
      description: Refused with `not_terminal` while a parse that has not ended reads it.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses:
    post:
      tags: [parses]
      operationId: createParse
      summary: Submit a parse
      description: |
        Answers `202` with the parse. With `Prefer: wait=<seconds>` the
        answer is held until the parse ends or the wait runs out, and is
        `200` when it ended.

        A submit always makes a parse of its own, with its own id, labels
        and origin. What is not repeated is the reading: a page that an
        earlier parse of the same caller read whole, from the same bytes
        with the same readers, is taken from that read and no model is
        called for it, unless `reuse` is false. Such a page says
        `reused: true` and the parse counts it in `progress.pages_reused`.
      parameters:
        - name: Idempotency-Key
          in: header
          description: Makes the submit safe to repeat for 24 hours. The same key with a different body is `idempotency_conflict`.
          schema: { type: string, maxLength: 255 }
        - name: Prefer
          in: header
          description: "`wait=<seconds>`, at most 60."
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ParseRequest" }
      responses:
        "200": { description: The parse ended within the wait., content: { application/json: { schema: { $ref: "#/components/schemas/Parse" } } } }
        "202": { description: Accepted., content: { application/json: { schema: { $ref: "#/components/schemas/Parse" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    get:
      tags: [parses]
      operationId: listParses
      summary: List parses, newest first
      parameters:
        - { name: state, in: query, schema: { $ref: "#/components/schemas/ParseState" } }
        - { name: file, in: query, schema: { type: string } }
        - { name: origin.path, in: query, description: Parses submitted with this origin path., schema: { type: string } }
        - { name: label, in: query, description: "`key=value`; may repeat, and every one must match.", schema: { type: array, items: { type: string } }, style: form, explode: true }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200": { description: A page of parses., content: { application/json: { schema: { $ref: "#/components/schemas/ParseList" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /parses/{parse}:
    parameters:
      - $ref: "#/components/parameters/parse"
    get:
      tags: [parses]
      operationId: getParse
      summary: Read a parse
      description: Read a parse's state, progress, and usage.
      responses:
        "200": { description: The parse., content: { application/json: { schema: { $ref: "#/components/schemas/Parse" } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [parses]
      operationId: deleteParse
      summary: Delete a parse
      description: |
        Delete a parse and everything it wrote.

        Refused with `not_terminal` for a parse that has not ended; cancel it first.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/cancel:
    parameters:
      - $ref: "#/components/parameters/parse"
    post:
      tags: [parses]
      operationId: cancelParse
      summary: Stop a parse
      description: |
        No further page is read, and a page being read is abandoned. What
        was already read stays readable. Canceling a parse that has ended
        is `already_terminal`.
      responses:
        "200": { description: The parse, canceled or on its way., content: { application/json: { schema: { $ref: "#/components/schemas/Parse" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/retry:
    parameters:
      - $ref: "#/components/parameters/parse"
    post:
      tags: [parses]
      operationId: retryParse
      x-lectio-status: planned
      summary: Retry failed pages
      description: |
        Read again the pages that failed.

        Pages that succeeded are not read again.
      responses:
        "202": { description: Queued again., content: { application/json: { schema: { $ref: "#/components/schemas/Parse" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
  /parses/{parse}/events:
    parameters:
      - $ref: "#/components/parameters/parse"
    get:
      tags: [parses]
      operationId: streamParse
      x-lectio-status: planned
      summary: Follow a parse
      description: |
        Follow a parse as server-sent events.

        `state`, `page` (one per page that ended, with its number and
        state), and `progress` events. Each carries an increasing id, and
        `Last-Event-ID` resumes after it.
      responses:
        "200": { description: The stream., content: { text/event-stream: { schema: { type: string } } } }
        "404": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
  /parses/{parse}/document:
    parameters:
      - $ref: "#/components/parameters/parse"
    get:
      tags: [results]
      operationId: getDocument
      summary: Read the document
      description: |
        As JSON this is the document's index: its pages without their
        blocks, the tables that continue across pages, the outline, and
        usage. As Markdown or text it is the content in reading order.
        Readable once the parse has assembled its pages, which a parse
        with failed pages still does.
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [json, markdown, text], default: json }
        - name: pages
          in: query
          description: Limits a Markdown or text rendering to these pages, in the syntax of a submit's `pages`.
          schema: { type: string }
        - name: tables
          in: query
          description: How a table is written in Markdown. `auto` writes Markdown unless a cell spans.
          schema: { type: string, enum: [auto, markdown, html], default: auto }
        - name: repeated
          in: query
          description: Whether running headers and footers are printed on every page or once.
          schema: { type: string, enum: [once, keep, drop], default: once }
        - name: page_breaks
          in: query
          description: Whether a Markdown rendering marks where each page begins.
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: The document.
          content:
            application/json: { schema: { $ref: "#/components/schemas/Document" } }
            text/markdown: { schema: { type: string } }
            text/plain: { schema: { type: string } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/pages:
    parameters:
      - $ref: "#/components/parameters/parse"
    get:
      tags: [results]
      operationId: listPages
      summary: List pages
      description: List the pages and their states, without blocks.
      responses:
        "200": { description: The pages., content: { application/json: { schema: { $ref: "#/components/schemas/PageList" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /parses/{parse}/pages/{page}:
    parameters:
      - $ref: "#/components/parameters/parse"
      - $ref: "#/components/parameters/page"
    get:
      tags: [results]
      operationId: getPage
      summary: Read one page
      description: |
        Read one page with its blocks.

        Readable as soon as that page has been read, while the parse is
        still running. A page not read yet is `409 page_not_ready`. A page
        the parse does not read, or one it ended without reading, is
        `404 page_not_found`.
      responses:
        "200": { description: The page., content: { application/json: { schema: { $ref: "#/components/schemas/Page" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/pages/{page}/image:
    parameters:
      - $ref: "#/components/parameters/parse"
      - $ref: "#/components/parameters/page"
    get:
      tags: [results]
      operationId: getPageImage
      summary: Read a page image
      description: |
        Read the image of a page that a reader saw.

        The image is the one sent to the reader, so a block's box drawn over it lines up. A page read natively has none.
      responses:
        "200":
          description: The image.
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/blocks:
    parameters:
      - $ref: "#/components/parameters/parse"
    get:
      tags: [results]
      operationId: listBlocks
      summary: Read every block
      description: |
        Read every block of the pages read so far.

        The bulk read: all blocks in page and reading order, one per line,
        in one answer. It answers while the parse runs, with the pages
        read so far. A block's `ref` says which page it is on.
      parameters:
        - name: pages
          in: query
          description: Limits the answer to these pages, in the syntax of a submit's `pages`.
          schema: { type: string }
      responses:
        "200": { description: One block per line., content: { application/x-ndjson: { schema: { $ref: "#/components/schemas/Block" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/blocks/{ref}:
    parameters:
      - $ref: "#/components/parameters/parse"
      - name: ref
        in: path
        required: true
        description: "`<page>.<order>`"
        schema: { type: string, pattern: '^[1-9][0-9]*\.[1-9][0-9]*$' }
    get:
      tags: [results]
      operationId: getBlock
      summary: Read one block
      responses:
        "200": { description: The block., content: { application/json: { schema: { $ref: "#/components/schemas/Block" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/blocks/{ref}/image:
    parameters:
      - $ref: "#/components/parameters/parse"
      - name: ref
        in: path
        required: true
        description: "`<page>.<order>`"
        schema: { type: string, pattern: '^[1-9][0-9]*\.[1-9][0-9]*$' }
    get:
      tags: [results]
      operationId: getBlockImage
      summary: Read a block image
      description: |
        Read a block as it looks on its page.

        The page's image cut to the block's box: a figure, a table, a
        signature as an image of its own. The page image is the one the
        reader saw, so the cut is the region the box names. A block with
        no box, and a block of a page that was not read from an image, is
        `404 block_not_found`.
      responses:
        "200":
          description: The image.
          content:
            image/png: { schema: { type: string, format: binary } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /parses/{parse}/figures:
    parameters:
      - $ref: "#/components/parameters/parse"
    post:
      tags: [figures]
      operationId: createFigures
      summary: Describe figures
      description: |
        Describe the figures of a parse.

        A reader that finds a figure says where it is. This asks a
        describer what each figure shows: the figure is cut from its
        page's image and described alone, and the answer is written onto
        the figure's block as `description`, with `figure.type` and, when
        the page's reader gave the block no text, the words printed in
        the figure as `text`.

        It is a request against a parse that has ended, so figures can be
        described later, for some pages, and no page is read again. Only
        a figure that has a box, on a page read from an image, and has no
        description yet is described, unless `redo` is true. A figure
        described once from the same bytes by the same describers is not
        described again by a later parse. One run at a time: a request
        while a run is in flight is `conflict`. A parse that has not
        ended is `not_terminal`. With no describer configured, or one
        named that is not, it is `reader_not_found`.

        Answers `202` with the run. With `Prefer: wait=<seconds>` the
        answer is held until the run ends or the wait runs out, and is
        `200` when it ended. A run succeeds when it described any figure
        it set out to; a figure it lost says why in the listing.

        A server that keeps its work in a database does not describe
        figures yet and answers `501 not_implemented`; a development
        server does.
      parameters:
        - name: Prefer
          in: header
          description: "`wait=<seconds>`, at most 60."
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FiguresRequest" }
      responses:
        "200": { description: The run ended within the wait., content: { application/json: { schema: { $ref: "#/components/schemas/Figures" } } } }
        "202": { description: Accepted., content: { application/json: { schema: { $ref: "#/components/schemas/Figures" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
    get:
      tags: [figures]
      operationId: listFigures
      summary: List figures
      description: |
        List a parse's figures and the run that describes them.

        Every figure block of the parse, described or not. `run` is absent when no run was started.
      responses:
        "200": { description: The figures., content: { application/json: { schema: { $ref: "#/components/schemas/Figures" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /parses/{parse}/fields/{name}:
    parameters:
      - $ref: "#/components/parameters/parse"
      - { name: name, in: path, required: true, schema: { type: string, pattern: '^[a-z][a-z0-9_-]{0,62}$' } }
    get:
      tags: [fields]
      operationId: getField
      x-lectio-status: planned
      summary: Read one extraction
      parameters:
        - name: resolve
          in: query
          description: Returns each citation with its page and box, and not as a ref alone.
          schema: { type: boolean, default: false }
      responses:
        "200": { description: The extraction., content: { application/json: { schema: { $ref: "#/components/schemas/Field" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    file:
      name: file
      in: path
      required: true
      schema: { type: string, pattern: '^fil_[0-9A-Z]{26}$' }
...